Moved `brew-install` to trenalstor class
